Update unit tests

+++ src/test/unit/flight_imu_unittest.cc
@@ -204,7 +204,7 @@ TEST(FlightImuTest, TestSmallAngle)
 extern "C" {
 boxBitmask_t rcModeActivationMask;
 float rcCommand[4];
-int16_t rcData[MAX_SUPPORTED_RC_CHANNEL_COUNT];
+float rcData[MAX_SUPPORTED_RC_CHANNEL_COUNT];
 
 gyro_t gyro;
 acc_t acc;
